release 3.0.57

+++ src/main/java/com/tencentcloudapi/vod/v20180717/models/AiAnalysisTaskCoverResult.java
@@ -35,7 +35,7 @@ public class AiAnalysisTaskCoverResult  extends AbstractModel{
     */
     @SerializedName("ErrCode")
     @Expose
-    private Integer ErrCode;
+    private Long ErrCode;
 
     /**
     * 错误信息。
@@ -82,7 +82,7 @@ public class AiAnalysisTaskCoverResult  extends AbstractModel{
      * @return ErrCode 错误码，0：成功，其他值：失败。
 注意：此字段可能返回 null，表示取不到有效值。
      */
-    public Integer getErrCode() {
+    public Long getErrCode() {
         return this.ErrCode;
     }
 
@@ -92,7 +92,7 @@ public class AiAnalysisTaskCoverResult  extends AbstractModel{
      * @param ErrCode 错误码，0：成功，其他值：失败。
 注意：此字段可能返回 null，表示取不到有效值。
      */
-    public void setErrCode(Integer ErrCode) {
+    public void setErrCode(Long ErrCode) {
         this.ErrCode = ErrCode;
     }
