+++ src/main/java/com/tencentcloudapi/vod/v20180717/models/AiAnalysisTaskTagInput.java
@@ -27,13 +27,13 @@ public class AiAnalysisTaskTagInput  extends AbstractModel{
     */
     @SerializedName("Definition")
     @Expose
-    private Integer Definition;
+    private Long Definition;
 
     /**
      * 获取视频智能分类模板 ID ，固定为 10 。
      * @return Definition 视频智能分类模板 ID ，固定为 10 。
      */
-    public Integer getDefinition() {
+    public Long getDefinition() {
         return this.Definition;
     }
 
@@ -41,7 +41,7 @@ public class AiAnalysisTaskTagInput  extends AbstractModel{
      * 设置视频智能分类模板 ID ，固定为 10 。
      * @param Definition 视频智能分类模板 ID ，固定为 10 。
      */
-    public void setDefinition(Integer Definition) {
+    public void setDefinition(Long Definition) {
         this.Definition = Definition;
     }
