+++ src/main/java/com/tencentcloudapi/vod/v20180717/models/CoverBySnapshotTaskInput.java
@@ -27,7 +27,7 @@ public class CoverBySnapshotTaskInput  extends AbstractModel{
     */
     @SerializedName("Definition")
     @Expose
-    private Integer Definition;
+    private Long Definition;
 
     /**
     * 截图方式。包含：
@@ -59,7 +59,7 @@ public class CoverBySnapshotTaskInput  extends AbstractModel{
      * 获取指定时间点截图模板 ID。
      * @return Definition 指定时间点截图模板 ID。
      */
-    public Integer getDefinition() {
+    public Long getDefinition() {
         return this.Definition;
     }
 
@@ -67,7 +67,7 @@ public class CoverBySnapshotTaskInput  extends AbstractModel{
      * 设置指定时间点截图模板 ID。
      * @param Definition 指定时间点截图模板 ID。
      */
-    public void setDefinition(Integer Definition) {
+    public void setDefinition(Long Definition) {
         this.Definition = Definition;
     }
