+++ src/main/java/com/tencentcloudapi/vod/v20180717/models/CreateClassRequest.java
@@ -27,7 +27,7 @@ public class CreateClassRequest  extends AbstractModel{
     */
     @SerializedName("ParentId")
     @Expose
-    private Integer ParentId;
+    private Long ParentId;
 
     /**
     * 分类名称，长度限制：1-64 个字符。
@@ -41,13 +41,13 @@ public class CreateClassRequest  extends AbstractModel{
     */
     @SerializedName("SubAppId")
     @Expose
-    private Integer SubAppId;
+    private Long SubAppId;
 
     /**
      * 获取父类 ID，一级分类填写 -1。
      * @return ParentId 父类 ID，一级分类填写 -1。
      */
-    public Integer getParentId() {
+    public Long getParentId() {
         return this.ParentId;
     }
 
@@ -55,7 +55,7 @@ public class CreateClassRequest  extends AbstractModel{
      * 设置父类 ID，一级分类填写 -1。
      * @param ParentId 父类 ID，一级分类填写 -1。
      */
-    public void setParentId(Integer ParentId) {
+    public void setParentId(Long ParentId) {
         this.ParentId = ParentId;
     }
 
@@ -79,7 +79,7 @@ public class CreateClassRequest  extends AbstractModel{
      * 获取点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
      * @return SubAppId 点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
      */
-    public Integer getSubAppId() {
+    public Long getSubAppId() {
         return this.SubAppId;
     }
 
@@ -87,7 +87,7 @@ public class CreateClassRequest  extends AbstractModel{
      * 设置点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
      * @param SubAppId 点播[子应用](/document/product/266/14574) ID。如果要访问子应用中的资源，则将该字段填写为子应用 ID；否则无需填写该字段。
      */
-    public void setSubAppId(Integer SubAppId) {
+    public void setSubAppId(Long SubAppId) {
         this.SubAppId = SubAppId;
     }
